+++ Tests/Fixtures/Bundle/Entity/DummySubject.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Sonata\AdminBundle\Tests\Fixtures\Bundle\Entity;
+
+class DummySubject
+{
+    public function __toString()
+    {
+        return 'dummy subject representation';
+    }
+}